Updated to Naut 4.0
Updated the images and CSS file to Naut 4.0, along with bugfixes and
tweaks to match.

This update changes the style a decent amount, so don't be shocked.

I've tweaked the h5 header to work in this version, that's the one that
goes above the posts, but the h6 header wasn't working so it's commented
out.

Some color tweaks were added near the bottom, feel free to change those.

The 'footer' and 'sidebarimg' images are default, so they don't really
suit the sub, so they will need to be updated to something you choose.

RES-Nightmode support is included at the bottom, so you can add the text
'[](#/RES_SR_Config/NightModeCompatible)' somewhere in the sidebar text
to have it enabled by default.

It shows as thousands of lines removed because they minified the Naut
theme to take up just a single line now FYI